convert latest vue.js
