+++ README.md
@@ -1,46 +1,29 @@
-# ballot-labo
+# new-app
 
-> vote app
-
-## Build Setup
-
-``` bash
-# install dependencies
+## Project setup
+```
 npm install
+```
 
-# serve with hot reload at localhost:8080
-npm run dev
+### Compiles and hot-reloads for development
+```
+npm run serve
+```
 
-# build for production with minification
+### Compiles and minifies for production
+```
 npm run build
-
-# build for production and view the bundle analyzer report
-npm run build --report
 ```
 
-For a detailed explanation on how things work, check out the [guide](http://vuejs-templates.github.io/webpack/) and [docs for vue-loader](http://vuejs.github.io/vue-loader).
-
-
-
+### Run your tests
 ```
-npm install -g firebase-tools
-npm install -g vue-cli
-
-view init webpack rhd2018-ballot
-cd rhd2018-ballot
-
-npm run start
-
-firebase login
-firebase init
-
-firebase serve
-firebase deploy
+npm run test
 ```
 
-## 追加ライブラリ
+### Lints and fixes files
 ```
-npm install --save-dev bulma sass-loader node-sass 
-npm install --save @fortawesome/fontawesome-free-webfonts
-npm install --save firebase vuefire
+npm run lint
 ```
+
+### Customize configuration
+See [Configuration Reference](https://cli.vuejs.org/config/).
