--- config/prod.env.js
@@ -1,4 +1,0 @@
-'use strict'
-module.exports = {
-  NODE_ENV: '"production"'
-}
